+++ flexiapi/resources/views/errors/490.blade.php
@@ -1,0 +1,11 @@
+@extends('layouts.main')
+
+@section('content')
+
+<h2><i class="ph">globe-hemisphere-west</i> {{ __('Expired Space') }}</h2>
+
+<p class="text-center">
+    {{ __('Your space has expired. Access to your interface is now disabled, and your users can no longer benefit from the service. To reactivate your space, please contact your account manager.') }}
+</p>
+
+@endsection